Add support to auto-update your .arcrc based on changes to the default .arcrc.
We add a new default khan-arcrc in khan-bin/khan-arcrc.  Whenever we
sync the new arcanist repo (once per day), we may also get a new copy
of the default khan-arcrc.  At that time, we check if the new
khan-arcrc differs from the one you have installed in ~/.arcrc, and if
so it updates it so they're consistent.

A new khan/do_not_auto_update field is added to suppress this behavior
on a field-by-field basis, and I set up a default value in the
khan-arcrc to provide an example of how to use it.

Tested: additions to arc_unittest, and manual testing on my personal
.arcrc.